move registry outside of lib
