+++ apps/www/scripts/build-registry.ts
@@ -8,7 +8,7 @@ import { colorMapping, colors } from "../src/lib/registry/colors";
 import { registrySchema } from "../src/lib/registry/schema";
 import { styles } from "../src/lib/registry/styles";
 import { themes } from "../src/lib/registry/themes";
-import { buildRegistry } from "../src/lib/registry/registry";
+import { buildRegistry } from "./registry";
 
 const REGISTRY_PATH = path.join(process.cwd(), "static/registry");
