+++ apps/www/scripts/registry.ts
@@ -1,7 +1,7 @@
 import { readdir, readFile } from "node:fs/promises";
 import { resolve, join } from "node:path";
 import { parse, preprocess, walk } from "svelte/compiler";
-import config from "../../../svelte.config.js";
+import config from "../svelte.config.js";
 import type { Registry } from "@/registry/schema";
 
 const DEPENDENCIES = ["@huntabyte/primitives"];
